+++ AssetTaggerDatabase/dbo/Tables/ProductSet/ProductSet.sql
@@ -2,10 +2,10 @@
     -- Non-nullable columns with default values.
     [CreatedAt] DATETIME2(3) CONSTRAINT [DF_ProductSet_CreatedAt] DEFAULT (SYSUTCDATETIME()) NOT NULL,
 
-    [ProductQuantity] INT DEFAULT 1 NOT NULL,
+    [ProductQuantity] BIGINT DEFAULT 1 NOT NULL,
     CONSTRAINT [CK_ProductSet_ProductQuantity] CHECK ([ProductQuantity] > 0),
 
-    [RowNumber] INT IDENTITY (1, 1),
+    [RowNumber] BIGINT IDENTITY (1, 1),
     CONSTRAINT [AK_ProductSet_RowNumber] UNIQUE CLUSTERED ([RowNumber]),
 
     -- Non-nullable foreign keys.
